Available room results now takes into account date ranges -- Read commit details carefully
Updates:
1. Added @NamedNativeQuery findAvailableRooms that queries both bookings
and package_bookings before returning a list of available rooms. Only
rooms not booked during the user's searched date range will be returned.
This is tested and appears to be working.


***Read carefully***
2. Changed data type of checkInDate and checkOutDate from String to
java.sql.Date in BOTH Booking and PackageBooking domain classes
This was required to perform a valid date comparison in MySQL

3. Changed data type of check_in_date and check_out_date from VARCHAR to
Date in bookings and package_bookings tables. Please DROP both tables in
your database and use the new table creation scripts in the SQL scripts
file in this commit.

4. Added method to convert String to java.sql.Date:
	NewBookingService.stringToSqlDate() -- use if you need to make this
conversion.

5. Fixed multiple tests which the data type change broke. Currently all
are passing.